requirements.txt
torch>=1.8.0
torchvision>=0.9.0
numpy>=1.19.0
matplotlib>=3.3.0
opencv-python-headless>=4.5.0
scikit-learn>=1.0.0
albumentations>=1.0.0
segmentation-models-pytorch>=0.3.0
tqdm>=4.60.0
Pillow>=8.0.0
scikit-image>=0.18.0
PyYAML>=5.1